Extract Attributes klass from Ranking klass
