+++ lib/dmm-crawler.rb
@@ -1,8 +1,10 @@
 require 'mechanize'
 
 module DMMCrawler
+  BASE_URL = 'http://www.dmm.co.jp'.freeze
 end
 
 require 'dmm-crawler/agent'
+require 'dmm-crawler/attributes'
 require 'dmm-crawler/ranking'
 require 'dmm-crawler/version'
